# backend/answer-service/src/main/java/com/microservices/answerservice/models/entity/Answer.java
package com.microservices.answerservice.models.entity;

import com.microservices.commonexam.models.entity.Question;
import com.microservices.commonstudent.models.entity.Student;
import lombok.Getter;
import lombok.Setter;
import org.springframework.data.annotation.Id;
import org.springframework.data.annotation.Transient;
import org.springframework.data.mongodb.core.mapping.Document;





@Getter
@Setter
@Document(collection = "answers")
public class Answer {

    @Id
    private String id;

    private String text;

    @Transient
    private Student student;

    private Long studentId;

    @Transient
    private Question question;

    private Long questionId;

    private Long examId; // Ajouter ce champ


    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public String getText() {
        return text;
    }

    public void setText(String text) {
        this.text = text;
    }

    public Student getStudent() {
        return student;
    }

    public void setStudent(Student student) {
        this.student = student;
    }

    public Long getStudentId() {
        return studentId;
    }

    public void setStudentId(Long studentId) {
        this.studentId = studentId;
    }

    public Question getQuestion() {
        return question;
    }

    public void setQuestion(Question question) {
        this.question = question;
    }

    public Long getQuestionId() {
        return questionId;
    }

    public void setQuestionId(Long questionId) {
        this.questionId = questionId;
    }

    public Long getExamId() {
        return examId;
    }

    public void setExamId(Long examId) {
        this.examId = examId;
    }



}